OpenTelemetry-Prometheus/basic-python-server/src/requirements.txt
Flask == 2.3.2
prometheus_client == 0.14.1
opentelemetry-api==1.20.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-sdk==1.20.0